requirements.txt
streamlit>=1.30
python-dotenv>=1.0
requests>=2.31
openai>=1.0
